requirements.txt
easydict
numpy
torch == 2.1.2
torchvision == 0.16.2
tqdm
pyyaml